move transcoding into Transcoder thread
Transcoding is now done by the Transcoder thread.  Updates are sent to the
GUI thread, which updates the status label.  The transcoding logic was
moved from Frontend::transcode() to Transcoder::run().

There is currently no way to nicely stop the transcoding process.  This
will be fixed in a later version.